refactor!: switch to thiserror (from failure)
