+++ src/interpolate.rs
@@ -2,23 +2,7 @@
 
 use crate::point::Point;
 use crate::source::Source;
-use failure::Fail;
-
-/// Errors for interpolation.
-#[derive(Clone, Copy, Debug, Fail)]
-pub enum Error {
-    /// Error returned when trying to extrapolate with only one point in the source.
-    #[fail(display = "Cannot interpolate in a source with only one point")]
-    OnePoint,
-
-    /// The time value is below the minimum time of the source.
-    #[fail(display = "Time value is below minimum of the source: {}", _0)]
-    TimeBelowMinimum(f64),
-
-    /// The time value is above the maximum time of the source.
-    #[fail(display = "Time value is above the maximum of the source: {}", _0)]
-    TimeAboveMaximum(f64),
-}
+use crate::Error;
 
 /// Structure that handles the interpolation.
 #[derive(Debug)]
@@ -39,7 +23,7 @@ impl Interpolator {
     /// let reader = sbet::Reader::from_path("data/2-points.sbet").unwrap();
     /// let interpolator = Interpolator::new(Box::new(reader)).unwrap();
     /// ```
-    pub fn new(mut source: Box<dyn Source>) -> Result<Interpolator, failure::Error> {
+    pub fn new(mut source: Box<dyn Source>) -> Result<Interpolator, Error> {
         let mut points = Vec::with_capacity(2);
         for _ in 0..2 {
             points.push(match source.source()? {
@@ -67,7 +51,7 @@ impl Interpolator {
     /// let mut interpolator = Interpolator::new(Box::new(reader)).unwrap();
     /// let point = interpolator.interpolate(1.516310048360710e5).unwrap();
     /// ```
-    pub fn interpolate(&mut self, time: f64) -> Result<Point, failure::Error> {
+    pub fn interpolate(&mut self, time: f64) -> Result<Point, Error> {
         loop {
             assert!(self.index != 0 && self.index != self.points.len());
             if time < self.points[self.index - 1].time {
