+++ src/pof.rs
@@ -5,26 +5,13 @@
 use crate::point::Point;
 use crate::source::Source;
 use crate::units::Radians;
+use crate::Error;
 use byteorder::{LittleEndian, ReadBytesExt};
-use failure;
-use failure::Fail;
 use std::fmt::Debug;
 use std::fs::File;
 use std::io::{BufReader, Read, Seek, SeekFrom};
 use std::iter::IntoIterator;
 use std::path::Path;
-
-/// Pof errors.
-#[derive(Clone, Copy, Debug, Fail)]
-pub enum Error {
-    /// The time unit code is invalid.
-    #[fail(display = "The time unit code is invalid: {}", _0)]
-    TimeUnit(u8),
-
-    /// The time info code is invalid.
-    #[fail(display = "The time info code is invalid: {}", _0)]
-    TimeInfo(u8),
-}
 
 /// A pos file reader.
 #[derive(Debug)]
@@ -108,14 +95,14 @@ impl Reader<BufReader<File>> {
     /// use pos::pof::Reader;
     /// let reader = Reader::from_path("data/sbet_mission_1.pof").unwrap();
     /// ```
-    pub fn from_path<P: AsRef<Path>>(path: P) -> Result<Reader<BufReader<File>>, failure::Error> {
+    pub fn from_path<P: AsRef<Path>>(path: P) -> Result<Reader<BufReader<File>>, Error> {
         let reader = BufReader::new(File::open(path)?);
         Reader::new(reader)
     }
 }
 
 impl<R: Read + Seek> Reader<R> {
-    fn new(mut reader: R) -> Result<Reader<R>, failure::Error> {
+    fn new(mut reader: R) -> Result<Reader<R>, Error> {
         let mut preamble = [0; 27];
         reader.read_exact(&mut preamble)?;
 
@@ -193,7 +180,7 @@ impl<R: Read + Seek> Reader<R> {
     /// let mut reader = Reader::from_path("data/sbet_mission_1.pof").unwrap();
     /// let point = reader.read_point().unwrap();
     /// ```
-    pub fn read_point(&mut self) -> Result<Option<Point>, failure::Error> {
+    pub fn read_point(&mut self) -> Result<Option<Point>, Error> {
         if self.position == self.entries {
             return Ok(None);
         }
@@ -290,7 +277,7 @@ impl TimeUnit {
             0 => Ok(TimeUnit::Normalized),
             1 => Ok(TimeUnit::Day),
             2 => Ok(TimeUnit::Week),
-            _ => Err(Error::TimeUnit(n)),
+            _ => Err(Error::PofTimeUnit(n)),
         }
     }
 }
@@ -312,13 +299,13 @@ impl TimeInfo {
             0 => Ok(TimeInfo::Gps),
             1 => Ok(TimeInfo::Utc),
             2 => Ok(TimeInfo::Unknown),
-            _ => Err(Error::TimeInfo(n)),
+            _ => Err(Error::PofTimeInfo(n)),
         }
     }
 }
 
 impl<R: Debug + Seek + Read> Source for Reader<R> {
-    fn source(&mut self) -> Result<Option<Point>, failure::Error> {
+    fn source(&mut self) -> Result<Option<Point>, Error> {
         self.read_point()
     }
 }
